+++ packages/python-ta/src/python_ta/reporters/html_reporter.py
@@ -44,7 +44,6 @@ class HTMLReporter(PythonTaReporter):
     }
     _PRE_LINE_NUM_SPACES = 0
 
-    no_err_message = "No problems detected, good job!"
     no_snippet = "No code to display for this message."
     code_err_title = "Code Errors or Forbidden Usage (fix: high priority)"
     style_err_title = "Style or Convention Errors (fix: before submission)"